Add vanilla CSS support to RadioGroup (#522)
