+++ components/demos/RadioGroup/css/styles.css
@@ -1,0 +1,52 @@
+@import '@radix-ui/colors/blackA.css';
+@import '@radix-ui/colors/violet.css';
+
+/* reset */
+button {
+  all: unset;
+}
+
+.RadioGroupRoot {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.RadioGroupItem {
+  background-color: white;
+  width: 25px;
+  height: 25px;
+  border-radius: 100%;
+  box-shadow: 0 2px 10px var(--blackA7);
+}
+.RadioGroupItem:hover {
+  background-color: var(--violet3);
+}
+.RadioGroupItem:focus {
+  box-shadow: 0 0 0 2px black;
+}
+
+.RadioGroupIndicator {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
+  position: relative;
+}
+.RadioGroupIndicator::after {
+  content: '';
+  display: block;
+  width: 11px;
+  height: 11px;
+  border-radius: 50%;
+  background-color: var(--violet11);
+}
+
+.Label {
+  color: white;
+  font-size: 15px;
+  line-height: 1;
+  user-select: none;
+  padding-left: 15px;
+}
