+++ data/primitives/components/radio-group/1.1.0.mdx
@@ -16,9 +16,7 @@ aria: https://www.w3.org/WAI/ARIA/apg/patterns/radiobutton
   <RadioGroupDemo />
 </HeroContainer>
 
-```jsx hero template=RadioGroup
-
-```
+<HeroCodeBlock folder="RadioGroup" />
 
 <Highlights
   features={[
